requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
cycler==0.10.0
dask==2.21.0
decorator==4.4.2
defusedxml==0.6.0
dlib==19.15.0
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.30.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.6
imageai==2.1.5
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
moviepy==1.0.3
mxnet==1.3.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.19.1
olefile==0.46
opencv-python==4.2.0.34
optuna==1.4.0
packaging==20.4
pafy==0.5.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pyttsx3==2.90
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.18.4
scikit-image==0.14.1
scikit-video==1.1.11
scipy==1.5.1
Send2Trash==1.5.0
six==1.15.0
sk-video==1.1.10
soupsieve==2.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.5.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
